refactor: cleanup of repo structure
