+++ .npmrc
@@ -1,0 +1,2 @@
+save-exact=true
+engine-strict=true
